semana5

+++ semana5.java
@@ -1,6 +1,6 @@
 public static semana5 {
     public static void main (String [] semana){
-        int fat = 1;
+        long fat = 1;
         for(int=5; i>=1;i--){
             fat=i*fat;
